+++ gpdb-20160503/alibabacloud_gpdb20160503/models/_reset_supabase_project_password_request.py
@@ -8,18 +8,19 @@ class ResetSupabaseProjectPasswordRequest(DaraModel):
     def __init__(
         self,
         account_password: str = None,
+        dashboard_password: str = None,
         project_id: str = None,
         region_id: str = None,
     ):
         # The password of the database account.
-        # 
-        # *   The password must contain at least three of the following character types: uppercase letters, lowercase letters, digits, and special characters.
-        # *   Special characters include `! @ # $ % ^ & * ( ) _ + - =`
-        # *   The password must be 8 to 32 characters in length.
+        # - The password must contain at least three of the following character types: uppercase letters, lowercase letters, digits, and special characters.
+        # - The following special characters are supported: `!@#$%^&*()_+-=`
+        # - The password must be 8 to 32 characters in length.
         # 
         # This parameter is required.
         self.account_password = account_password
-        # Supabase Instance ID
+        self.dashboard_password = dashboard_password
+        # The Supabase instance ID.
         # 
         # This parameter is required.
         self.project_id = project_id
@@ -37,6 +38,9 @@ class ResetSupabaseProjectPasswordRequest(DaraModel):
         if self.account_password is not None:
             result['AccountPassword'] = self.account_password
 
+        if self.dashboard_password is not None:
+            result['DashboardPassword'] = self.dashboard_password
+
         if self.project_id is not None:
             result['ProjectId'] = self.project_id
 
@@ -49,6 +53,9 @@ class ResetSupabaseProjectPasswordRequest(DaraModel):
         m = m or dict()
         if m.get('AccountPassword') is not None:
             self.account_password = m.get('AccountPassword')
+
+        if m.get('DashboardPassword') is not None:
+            self.dashboard_password = m.get('DashboardPassword')
 
         if m.get('ProjectId') is not None:
             self.project_id = m.get('ProjectId')